not expose NLogicalCores
Only needed for TPoolExecutor, no need for a user facing interface
