+++ core/base/inc/TROOT.h
@@ -97,7 +97,6 @@ namespace ROOT {
    void DisableImplicitMT();
    Bool_t IsImplicitMTEnabled();
    UInt_t GetImplicitMTPoolSize();
-   UInt_t NLogicalCores();
 }
 
 class TROOT : public TDirectory {
